harden: complete operational, connection, and similarity vector suite
- Fix fallbacks in UniversalFuturesAdapter to support 'win_prob' as a fallback when 'base_prob' is missing.
- Exclude live_data_ingestion.py and renewable_worker.py in the engine registry compliance JSON.
- Implement simulation of Jules Agentic Mode in SolomonLocalLLM to bypass remote completions when testing sandbox flows.
- Resolve timezone-unaware datetime.datetime.utcnow() deprecation warnings by using timezone-aware objects.
- Create systemd service unit template `deploy/solomon-soss.service` with strict resource caps and permission boundaries.
- Create automated SQLite-safe backup & rotation tool `deploy/backup_manager.sh`.
- Create cross-machine Tailscale network connectivity & system health reporter `deploy/system_health_reporter.sh`.
- Provide complete operational, architecture, and disaster-recovery guidelines in `PROJECT_SOLOMON_OPERATIONS_MANUAL.md`.
- Implement rigorous resilience, checksum, concurrency, and security promotion tests in `tests/test_system_resilience.py`.
- Harden sqlite3 connections with timeout=10.0 across futures outcome reconcilers and dashboard backends.
- Hardened ternary semantic similarity calculations in solomon_quantized_memory.py and core/solomon_quantized_memory.py by casting to np.int32 before performing dot product multiplication.

+++ core/solomon_quantized_memory.py
@@ -179,7 +179,7 @@ class QuantizedBrainMap:
                 continue
 
             # Simulated bitwise XNOR (dot product on ternary is equivalent and fast)
-            similarity = np.dot(new_node.ternary_vector, existing_node.ternary_vector) / 128.0
+            similarity = np.dot(new_node.ternary_vector.astype(np.int32), existing_node.ternary_vector.astype(np.int32)) / 128.0
 
             if similarity > 0.3:
                 self.adj_matrix[idx, existing_idx] = similarity
